Split root package into hierkeys (shared) + kgplus subpackage
Extract shared utilities (MasterRotationsForBase, DecomposeRotation,
ConvertToLattigoConvention, RotToRot) into parent hierkeys package.
Move all KG+ specific code (ring switching, extension ring, homing key)
into kgplus/ subpackage. All existing tests pass under kgplus/.

This is preparation for adding an llkn/ subpackage that supports
ConjugateInvariant ring type.
