+++ kgplus/derive.go
@@ -1,9 +1,10 @@
-package hierkeys
+package kgplus
 
 import (
 	"fmt"
 	"sort"
 
+	hierkeys "github.com/butvinm/lattigo-hierkeys"
 	"github.com/tuneinsight/lattigo/v6/core/rlwe"
 	"github.com/tuneinsight/lattigo/v6/ring"
 )
@@ -24,6 +25,45 @@ func DeriveGaloisKeys(params Parameters, tk *TransmissionKeys, targetRotations [
 	return eval.DeriveGaloisKeys(tk, targetRotations)
 }
 
+// RotToRot is a convenience wrapper that creates a temporary Evaluator
+// internally. For repeated calls, use [Evaluator.RotToRot].
+func RotToRot(
+	paramsLow rlwe.Parameters,
+	paramsHigh rlwe.Parameters,
+	inputKey *rlwe.GaloisKey,
+	masterKey *rlwe.GaloisKey,
+	combinedGalEl uint64,
+) (*rlwe.GaloisKey, error) {
+	params := Parameters{
+		Eval:         paramsLow, // placeholder
+		HK:           paramsLow, // placeholder
+		RPrime:       paramsLow,
+		RPrimeMaster: paramsHigh,
+	}
+	eval := NewEvaluator(params)
+	return eval.RotToRot(inputKey, masterKey, combinedGalEl)
+}
+
+// RingSwitchGaloisKey is a convenience wrapper that creates a temporary
+// Evaluator internally. For repeated calls, use [Evaluator.RingSwitchGaloisKey].
+func RingSwitchGaloisKey(
+	paramsEval rlwe.Parameters,
+	paramsHK rlwe.Parameters,
+	paramsRPrime rlwe.Parameters,
+	masterKeyRPrime *rlwe.GaloisKey,
+	homingKey *rlwe.EvaluationKey,
+	galoisElement uint64,
+) (*rlwe.GaloisKey, error) {
+	params := Parameters{
+		Eval:         paramsEval,
+		HK:           paramsHK,
+		RPrime:       paramsRPrime,
+		RPrimeMaster: paramsHK, // placeholder
+	}
+	eval := NewEvaluator(params)
+	return eval.RingSwitchGaloisKey(masterKeyRPrime, homingKey, galoisElement)
+}
+
 // DeriveGaloisKeys derives standard evaluation-level GaloisKeys from
 // transmission keys. The returned keys work with lattigo's standard
 // rlwe.Evaluator.Automorphism and ckks.Evaluator.Rotate.
@@ -31,9 +71,6 @@ func DeriveGaloisKeys(params Parameters, tk *TransmissionKeys, targetRotations [
 // This is a convenience wrapper that calls [Evaluator.ExpandInRPrime]
 // followed by [Evaluator.FinalizeKeys]. For finer control (e.g., storing
 // intermediate keys for later finalization), call those methods directly.
-//
-// The returned MemEvaluationKeySet can be passed directly to
-// rlwe.NewEvaluator or ckks.NewEvaluator.
 func (eval *Evaluator) DeriveGaloisKeys(tk *TransmissionKeys, targetRotations []int) (*rlwe.MemEvaluationKeySet, error) {
 
 	intermediate, err := eval.ExpandInRPrime(tk, targetRotations)
@@ -83,7 +120,7 @@ func (eval *Evaluator) ExpandInRPrime(tk *TransmissionKeys, targetRotations []in
 			continue // already computed
 		}
 
-		steps := decomposeRotation(normalized, masterRots)
+		steps := hierkeys.DecomposeRotation(normalized, masterRots)
 		if steps == nil {
 			return nil, fmt.Errorf("cannot decompose rotation %d (normalized from %d) from available masters",
 				normalized, target)
@@ -108,8 +145,7 @@ func (eval *Evaluator) ExpandInRPrime(tk *TransmissionKeys, targetRotations []in
 		}
 	}
 
-	// Extract requested targets (use original rotation index as key,
-	// normalized index for cache lookup)
+	// Extract requested targets
 	result := &IntermediateKeys{Keys: make(map[int]*rlwe.GaloisKey, len(targetRotations))}
 	for _, target := range targetRotations {
 		normalized := ((target % nSlots) + nSlots) % nSlots
@@ -157,8 +193,155 @@ func (eval *Evaluator) FinalizeKeys(tk *TransmissionKeys, intermediate *Intermed
 	return rlwe.NewMemEvaluationKeySet(nil, galoisKeys...), nil
 }
 
+// RingSwitchGaloisKey ring-switches a GaloisKey from R' (degree 2N) to a
+// standard GaloisKey in R (degree N) using a homing key.
+func (eval *Evaluator) RingSwitchGaloisKey(
+	masterKeyRPrime *rlwe.GaloisKey,
+	homingKey *rlwe.EvaluationKey,
+	galoisElement uint64,
+) (*rlwe.GaloisKey, error) {
+
+	paramsEval := eval.params.Eval
+	paramsHK := eval.params.HK
+	paramsRPrime := eval.params.RPrime
+
+	// Input validation
+	if paramsRPrime.N() != 2*paramsEval.N() {
+		return nil, fmt.Errorf("paramsRPrime.N()=%d must be 2*paramsEval.N()=%d", paramsRPrime.N(), 2*paramsEval.N())
+	}
+	if paramsHK.N() != paramsEval.N() {
+		return nil, fmt.Errorf("paramsHK.N()=%d must equal paramsEval.N()=%d", paramsHK.N(), paramsEval.N())
+	}
+	if paramsHK.QCount() != paramsEval.QCount()+paramsEval.PCount() {
+		return nil, fmt.Errorf("paramsHK.QCount()=%d must equal paramsEval.QCount()+PCount()=%d",
+			paramsHK.QCount(), paramsEval.QCount()+paramsEval.PCount())
+	}
+	if masterKeyRPrime == nil || homingKey == nil {
+		return nil, fmt.Errorf("masterKeyRPrime and homingKey must not be nil")
+	}
+
+	N := paramsEval.N()
+	ringQHK := paramsHK.RingQ()
+
+	levelQHK := paramsHK.MaxLevel()
+	levelQEval := paramsEval.MaxLevel()
+	levelPEval := paramsEval.MaxLevelP()
+
+	gc := &masterKeyRPrime.GadgetCiphertext
+	nRNS := len(gc.Value)
+
+	nEvalRNS := paramsEval.BaseRNSDecompositionVectorSize(levelQEval, levelPEval)
+	if nRNS < nEvalRNS {
+		return nil, fmt.Errorf("master key has %d RNS components, need at least %d", nRNS, nEvalRNS)
+	}
+
+	targetGK := &rlwe.GaloisKey{
+		EvaluationKey: rlwe.EvaluationKey{
+			GadgetCiphertext: *rlwe.NewGadgetCiphertext(
+				paramsEval, 1, levelQEval, levelPEval, 0),
+		},
+		GaloisElement: galoisElement,
+		NthRoot:       paramsEval.RingQ().NthRoot(),
+	}
+
+	ringQRPrimeQ := paramsRPrime.RingQ()
+	ringPRPrime := paramsRPrime.RingP()
+	ringQEval := paramsEval.RingQ()
+	ringPEval := paramsEval.RingP()
+
+	bQRPrime := eval.bQRPrime
+	aQRPrime := eval.aQRPrime
+	bQCoeff := eval.bQCoeff
+	aQCoeff := eval.aQCoeff
+	bPRPrime := eval.bPRPrime
+	aPRPrime := eval.aPRPrime
+	bPCoeff := eval.bPCoeff
+	aPCoeff := eval.aPCoeff
+	b0 := eval.b0RS
+	a0 := eval.a0RS
+	a1 := eval.a1RS
+	Xa1 := eval.Xa1RS
+	rsB := eval.rsBRS
+	rsA := eval.rsARS
+	ctKS := eval.ctKSRS
+
+	pIdx := levelQEval + 1 // start index of P_eval primes in Q_hk
+
+	for i := 0; i < nEvalRNS; i++ {
+		component := gc.Value[i][0]
+
+		// --- Extract even/odd from Q and P parts of R' component ---
+		bQRPrime.CopyLvl(paramsRPrime.MaxLevel(), component[0].Q)
+		aQRPrime.CopyLvl(paramsRPrime.MaxLevel(), component[1].Q)
+		ringQRPrimeQ.IMForm(bQRPrime, bQRPrime)
+		ringQRPrimeQ.IMForm(aQRPrime, aQRPrime)
+		ringQRPrimeQ.INTT(bQRPrime, bQCoeff)
+		ringQRPrimeQ.INTT(aQRPrime, aQCoeff)
+
+		// P parts
+		bPRPrime.CopyLvl(paramsRPrime.MaxLevelP(), component[0].P)
+		aPRPrime.CopyLvl(paramsRPrime.MaxLevelP(), component[1].P)
+		ringPRPrime.IMForm(bPRPrime, bPRPrime)
+		ringPRPrime.IMForm(aPRPrime, aPRPrime)
+		ringPRPrime.INTT(bPRPrime, bPCoeff)
+		ringPRPrime.INTT(aPRPrime, aPCoeff)
+
+		// Even/odd extraction into Q_hk-level polynomials
+		for m := 0; m <= paramsRPrime.MaxLevel(); m++ {
+			for j := 0; j < N; j++ {
+				b0.Coeffs[m][j] = bQCoeff.Coeffs[m][2*j]
+				a0.Coeffs[m][j] = aQCoeff.Coeffs[m][2*j]
+				a1.Coeffs[m][j] = aQCoeff.Coeffs[m][2*j+1]
+			}
+		}
+		for m := 0; m <= paramsRPrime.MaxLevelP(); m++ {
+			for j := 0; j < N; j++ {
+				b0.Coeffs[pIdx+m][j] = bPCoeff.Coeffs[m][2*j]
+				a0.Coeffs[pIdx+m][j] = aPCoeff.Coeffs[m][2*j]
+				a1.Coeffs[pIdx+m][j] = aPCoeff.Coeffs[m][2*j+1]
+			}
+		}
+
+		// Multiply a1 by X: X*f(X) mod (X^N+1)
+		for m := range a1.Coeffs {
+			qi := ringQHK.SubRings[m].Modulus
+			Xa1.Coeffs[m][0] = qi - a1.Coeffs[m][N-1]
+			for k := 1; k < N; k++ {
+				Xa1.Coeffs[m][k] = a1.Coeffs[m][k-1]
+			}
+		}
+
+		ringQHK.NTT(b0, b0)
+		ringQHK.NTT(a0, a0)
+		ringQHK.NTT(Xa1, Xa1)
+
+		// --- Key-switch X*a1 with homing key ---
+		eval.evalHK.GadgetProduct(levelQHK, Xa1, &homingKey.GadgetCiphertext, ctKS)
+
+		// Ring-switched ciphertext at Q_hk level
+		ringQHK.Add(b0, ctKS.Value[0], rsB)
+		ringQHK.Add(a0, ctKS.Value[1], rsA)
+
+		// --- Split Q_hk into Q_eval and P_eval parts ---
+		for m := 0; m <= levelQEval; m++ {
+			s := ringQEval.SubRings[m]
+			s.MForm(rsB.Coeffs[m], targetGK.Value[i][0][0].Q.Coeffs[m])
+			s.MForm(rsA.Coeffs[m], targetGK.Value[i][0][1].Q.Coeffs[m])
+		}
+		for m := 0; m <= levelPEval; m++ {
+			s := ringPEval.SubRings[m]
+			srcIdx := pIdx + m
+			s.MForm(rsB.Coeffs[srcIdx], targetGK.Value[i][0][0].P.Coeffs[m])
+			s.MForm(rsA.Coeffs[srcIdx], targetGK.Value[i][0][1].P.Coeffs[m])
+		}
+	}
+
+	return targetGK, nil
+}
+
 // convertToLattigoConvention applies pi^{-1} to each GadgetCiphertext component,
 // converting from paper convention to lattigo convention in-place.
+// Uses pre-allocated buffers for efficiency.
 func (eval *Evaluator) convertToLattigoConvention(gk *rlwe.GaloisKey) error {
 
 	paramsEval := eval.params.Eval
@@ -182,18 +365,15 @@ func (eval *Evaluator) convertToLattigoConvention(gk *rlwe.GaloisKey) error {
 		}
 	}
 
-	// Apply pi^{-1} to each component using pre-allocated buffers
 	for i := range gk.Value {
 		for j := range gk.Value[i] {
 			component := gk.Value[i][j]
 
-			// b component
 			eval.automorphInPlaceQ(ringQ, indexQ, component[0].Q)
 			if ringP != nil {
 				eval.automorphInPlaceP(ringP, indexP, component[0].P)
 			}
 
-			// a component
 			eval.automorphInPlaceQ(ringQ, indexQ, component[1].Q)
 			if ringP != nil {
 				eval.automorphInPlaceP(ringP, indexP, component[1].P)
@@ -204,68 +384,12 @@ func (eval *Evaluator) convertToLattigoConvention(gk *rlwe.GaloisKey) error {
 	return nil
 }
 
-// automorphInPlaceQ applies an automorphism to a polynomial at Q level
-// using the pre-allocated Q temporary buffer.
 func (eval *Evaluator) automorphInPlaceQ(r *ring.Ring, index []uint64, p ring.Poly) {
 	r.AutomorphismNTTWithIndex(p, index, eval.autTmpQ)
 	p.CopyLvl(p.Level(), eval.autTmpQ)
 }
 
-// automorphInPlaceP applies an automorphism to a polynomial at P level
-// using the pre-allocated P temporary buffer.
 func (eval *Evaluator) automorphInPlaceP(r *ring.Ring, index []uint64, p ring.Poly) {
 	r.AutomorphismNTTWithIndex(p, index, eval.autTmpP)
 	p.CopyLvl(p.Level(), eval.autTmpP)
-}
-
-// automorphInPlace applies an automorphism to a polynomial using a pre-computed
-// index. Allocates a temporary buffer internally — kept for backward compatibility
-// with tests that call convertToLattigoConvention as a package-level function.
-func automorphInPlace(r *ring.Ring, index []uint64, p ring.Poly) {
-	tmp := r.NewPoly()
-	r.AutomorphismNTTWithIndex(p, index, tmp)
-	p.CopyLvl(p.Level(), tmp)
-}
-
-// convertToLattigoConvention is the package-level backward-compatible wrapper.
-// It allocates temporary buffers per call. For repeated use, prefer
-// [Evaluator.convertToLattigoConvention].
-func convertToLattigoConvention(paramsEval rlwe.Parameters, gk *rlwe.GaloisKey) error {
-
-	galEl := gk.GaloisElement
-	galElInv := paramsEval.ModInvGaloisElement(galEl)
-
-	ringQ := paramsEval.RingQ()
-	ringP := paramsEval.RingP()
-
-	indexQ, err := ring.AutomorphismNTTIndex(ringQ.N(), ringQ.NthRoot(), galElInv)
-	if err != nil {
-		return fmt.Errorf("Q automorphism index: %w", err)
-	}
-
-	var indexP []uint64
-	if ringP != nil {
-		indexP, err = ring.AutomorphismNTTIndex(ringP.N(), ringP.NthRoot(), galElInv)
-		if err != nil {
-			return fmt.Errorf("P automorphism index: %w", err)
-		}
-	}
-
-	for i := range gk.Value {
-		for j := range gk.Value[i] {
-			component := gk.Value[i][j]
-
-			automorphInPlace(ringQ, indexQ, component[0].Q)
-			if ringP != nil {
-				automorphInPlace(ringP, indexP, component[0].P)
-			}
-
-			automorphInPlace(ringQ, indexQ, component[1].Q)
-			if ringP != nil {
-				automorphInPlace(ringP, indexP, component[1].P)
-			}
-		}
-	}
-
-	return nil
 }
