+++ kgplus/evaluator.go
@@ -1,6 +1,7 @@
-package hierkeys
+package kgplus
 
 import (
+	hierkeys "github.com/butvinm/lattigo-hierkeys"
 	"github.com/tuneinsight/lattigo/v6/core/rlwe"
 	"github.com/tuneinsight/lattigo/v6/ring"
 )
@@ -18,45 +19,32 @@ type Evaluator struct {
 }
 
 type evaluatorBuffers struct {
-	// --- Buffers for RingSwitchGaloisKey (eval.go) ---
+	// --- Buffers for RingSwitchGaloisKey ---
 
 	// R' Q-domain working space (degree 2N)
 	bQRPrime, aQRPrime ring.Poly // copy of component Q part
 	bQCoeff, aQCoeff   ring.Poly // after IMForm+INTT
-
 	// R' P-domain working space (degree 2N)
 	bPRPrime, aPRPrime ring.Poly // copy of component P part
 	bPCoeff, aPCoeff   ring.Poly // after IMForm+INTT
-
 	// Extracted even/odd at Q_hk level (degree N)
 	b0RS, a0RS, a1RS ring.Poly
 	Xa1RS            ring.Poly // X * a1
-
 	// Key-switch output and result at Q_hk level
 	rsBRS, rsARS ring.Poly
 	ctKSRS       *rlwe.Ciphertext
 	evalHK       *rlwe.Evaluator // evaluator at HK level for GadgetProduct
 
-	// --- Buffers for RotToRot (rottorot.go) ---
+	// --- Shared RotToRot buffers ---
+	rotBuf *hierkeys.RotToRotBuffers
 
-	// Combined Q+P at Q_hk level (degree 2N, RPrimeMaster ring)
-	bCombined, aCombined ring.Poly
-	// Automorphed at Q_hk level
-	bAut, aAut ring.Poly
-	// Key-switch output
-	ctKSRot *rlwe.Ciphertext
-	evalRot *rlwe.Evaluator // evaluator at RPrimeMaster level
-
-	// --- Buffers for convertToLattigoConvention / automorphInPlace ---
-
-	// Temporary poly for in-place automorphism (Q and P levels)
+	// --- Buffers for convertToLattigoConvention ---
 	autTmpQ ring.Poly // at eval Q level
 	autTmpP ring.Poly // at eval P level
 }
 
 // NewEvaluator creates an Evaluator with pre-allocated buffers for the
-// given hierarchical parameters. All subsequent calls to RingSwitchGaloisKey,
-// RotToRot, and DeriveGaloisKeys reuse these buffers.
+// given hierarchical parameters.
 func NewEvaluator(params Parameters) *Evaluator {
 	return &Evaluator{
 		params:           params,
@@ -68,7 +56,6 @@ func newEvaluatorBuffers(params Parameters) *evaluatorBuffers {
 	ringQRPrime := params.RPrime.RingQ()
 	ringPRPrime := params.RPrime.RingP()
 	ringQHK := params.HK.RingQ()
-	ringQRPMaster := params.RPrimeMaster.RingQ()
 	ringQEval := params.Eval.RingQ()
 
 	buf := &evaluatorBuffers{
@@ -90,20 +77,14 @@ func newEvaluatorBuffers(params Parameters) *evaluatorBuffers {
 		ctKSRS:   rlwe.NewCiphertext(params.HK, 1, params.HK.MaxLevel()),
 		evalHK:   rlwe.NewEvaluator(params.HK, nil),
 
-		// RotToRot buffers
-		bCombined: ringQRPMaster.NewPoly(),
-		aCombined: ringQRPMaster.NewPoly(),
-		bAut:      ringQRPMaster.NewPoly(),
-		aAut:      ringQRPMaster.NewPoly(),
-		ctKSRot:   rlwe.NewCiphertext(params.RPrimeMaster, 1, params.RPrimeMaster.MaxLevel()),
-		evalRot:   rlwe.NewEvaluator(params.RPrimeMaster, nil),
+		// Shared RotToRot buffers
+		rotBuf: hierkeys.NewRotToRotBuffers(params.RPrime, params.RPrimeMaster),
 
 		// automorphInPlace buffers
 		autTmpQ: ringQEval.NewPoly(),
 	}
 
 	buf.ctKSRS.IsNTT = true
-	buf.ctKSRot.IsNTT = true
 
 	if params.Eval.RingP() != nil {
 		buf.autTmpP = params.Eval.RingP().NewPoly()
@@ -113,11 +94,20 @@ func newEvaluatorBuffers(params Parameters) *evaluatorBuffers {
 }
 
 // ConcurrentCopy creates a copy of this Evaluator that shares read-only
-// data (parameters) but has its own mutable buffers. The original and
-// the copy can be used concurrently.
+// data (parameters) but has its own mutable buffers.
 func (eval *Evaluator) ConcurrentCopy() *Evaluator {
 	return &Evaluator{
 		params:           eval.params,
 		evaluatorBuffers: newEvaluatorBuffers(eval.params),
 	}
+}
+
+// RotToRot generates a combined rotation key from a level-0 key and a master
+// key in the extension ring R'. See [hierkeys.RotToRot] for details.
+func (eval *Evaluator) RotToRot(
+	inputKey *rlwe.GaloisKey,
+	masterKey *rlwe.GaloisKey,
+	combinedGalEl uint64,
+) (*rlwe.GaloisKey, error) {
+	return hierkeys.RotToRot(eval.rotBuf, eval.params.RPrime, eval.params.RPrimeMaster, inputKey, masterKey, combinedGalEl)
 }
