+++ kgplus/keygen.go
@@ -1,4 +1,4 @@
-package hierkeys
+package kgplus
 
 import (
 	"fmt"
@@ -70,7 +70,7 @@ func (kgen *KeyGenerator) ProjectToEvalKey(skHK *rlwe.SecretKey) *rlwe.SecretKey
 //
 // The sk must be generated by [GenSecretKeyNew] (at HK level).
 // masterRotations specifies which rotation indices to generate master keys for.
-// Use [MasterRotationsForBase] to compute a good set.
+// Use [hierkeys.MasterRotationsForBase] to compute a good set.
 //
 // The returned TransmissionKeys is sent to the server.
 func (kgen *KeyGenerator) GenTransmissionKeys(sk *rlwe.SecretKey, masterRotations []int) (*TransmissionKeys, error) {
@@ -146,7 +146,6 @@ func ConstructExtendedSK(paramsHK, paramsRP rlwe.Parameters, skS, skS1 *rlwe.Sec
 	skTilde := rlwe.NewSecretKey(paramsRP)
 
 	// Convert s and s̃₁ from NTT+Montgomery to coefficient domain.
-	// Both are at HK level (Q_eval ∪ P_eval primes).
 	sCoeffs := ringQHK.NewPoly()
 	s1Coeffs := ringQHK.NewPoly()
 	ringQHK.IMForm(skS.Value.Q, sCoeffs)
@@ -155,7 +154,6 @@ func ConstructExtendedSK(paramsHK, paramsRP rlwe.Parameters, skS, skS1 *rlwe.Sec
 	ringQHK.INTT(s1Coeffs, s1Coeffs)
 
 	// Interleave into R' (degree 2N): even = s, odd = s̃₁
-	// Use min(paramsRP.QCount(), paramsHK.QCount()) primes from HK coefficients
 	sTildeCoeffs := ringQRP.NewPoly()
 	nQ := paramsRP.QCount()
 	if nQ > paramsHK.QCount() {
